+++ ui/packages/shared/test-utils/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.0.19](https://github.com/parca-dev/parca/compare/@parca/test-utils@0.0.18...@parca/test-utils@0.0.19) (2025-12-18)
+
+**Note:** Version bump only for package @parca/test-utils
+
+## 0.0.18 (2025-12-18)
+
+**Note:** Version bump only for package @parca/test-utils
+
 ## 0.0.17 (2025-10-21)
 
 **Note:** Version bump only for package @parca/test-utils